+++ spec/gruf/rpc/service_spec.rb
@@ -56,10 +56,9 @@ describe Gruf::Service do
     end
   end
 
-  describe '.fail!' do
-    context 'on a failure' do
+  describe 'exceptions' do
+    context 'failing with a NotFound error' do
       subject { endpoint.get_fail(req, active_call) }
-      let(:metadata) { {foo: :bar} }
       let(:error) { Gruf::Error.new(code: :not_found, app_code: :thing_not_found, message: "#{req.id} not found!") }
 
       it 'should raise a GRPC::NotFound error' do
@@ -72,12 +71,47 @@ describe Gruf::Service do
       end
     end
 
+    context 'on an uncaught exception' do
+      subject { endpoint.get_uncaught_exception(req, active_call) }
+      let(:error) { Gruf::Error.new(code: :internal, app_code: :unknown, message: 'epic fail') }
+
+      it 'should raise a GRPC::Internal error' do
+        expect { subject }.to raise_error { |err|
+          expect(err).to be_a(GRPC::Internal)
+          expect(err.code).to eq 13
+          expect(err.message).to eq '13:epic fail'
+          expect(err.metadata).to eq(:'error-internal-bin' => error.serialize)
+        }
+      end
+
+      it 'should attach a backtrace if configured to do so' do
+        Gruf.backtrace_on_error = true
+        expect { subject }.to raise_error { |err|
+          error = JSON.parse(err.metadata[:'error-internal-bin'])
+          expect(error['debug_info']).to_not be_empty
+          expect(error['debug_info']['stack_trace']).to_not be_empty
+        }
+      end
+    end
+
     context 'on a success' do
       subject { endpoint.get_thing(req, active_call) }
 
       it 'should return normally' do
         expect(subject).to be_a(Rpc::GetThingResponse)
       end
+    end
+  end
+
+  describe '.method_added' do
+    it 'should add wrapper method for endpoint' do
+      expect(endpoint.respond_to?(:get_thing_with_intercept)).to be_truthy
+      expect(endpoint.respond_to?(:get_thing_without_intercept)).to be_truthy
+    end
+
+    it 'should not add wrapper method for non-endpoints' do
+      expect(endpoint.respond_to?(:not_a_endpoint_with_intercept)).to_not be_truthy
+      expect(endpoint.respond_to?(:not_a_endpoint_without_intercept)).to_not be_truthy
     end
   end